requirements.txt
beautifulsoup4==4.9.1
pyspellchecker==0.5.4
soupsieve==2.0.1